+++ src/cpu/x64/jit_uni_softmax.cpp
@@ -122,7 +122,7 @@ struct jit_softmax_dense_kernel_t : jit_softmax_kernel_base_t,
     size_t unroll_regs_ = 4;
 
     size_t axis_simd_full_;
-    size_t axis_simd_tail_;
+    int axis_simd_tail_;
     size_t n_loops_;
     size_t loop_tail_;
     size_t process_n_elems_;
@@ -959,7 +959,7 @@ struct jit_softmax_dense_kernel_t : jit_softmax_kernel_base_t,
                     tmp_vmm_injector, this->r14, this->r15, this->r13,
                     preserve_gpr, preserve_vmm,
                     PARAM_OFF(post_ops_binary_rhs_arg_vec), PARAM_OFF(dst_orig),
-                    dst_d_, static_cast<int>(axis_simd_tail_), tail_opmask,
+                    dst_d_, axis_simd_tail_, tail_opmask,
                     use_exact_tail_scalar_bcast};
 
             const binary_injector::static_params_t bsp {
@@ -1010,7 +1010,7 @@ struct jit_softmax_dense_kernel_t : jit_softmax_kernel_base_t,
                                   accumulation_mode::any)))
         , use_ext_aux_vmms_(!is_logsoftmax_ && n_vregs > 16)
         , axis_simd_full_(pd_->axis_size() / simd_w_)
-        , axis_simd_tail_(pd_->axis_size() % simd_w_) {
+        , axis_simd_tail_(static_cast<int>(pd_->axis_size() % simd_w_)) {
 
         const auto &post_ops = pd_->attr()->post_ops_;
         with_postops_ = post_ops.len() != 0;
@@ -1108,7 +1108,7 @@ struct jit_softmax_strided_kernel_t : jit_softmax_kernel_base_t,
     size_t axis_size_unroll_tail_;
     size_t axis_stride_;
     size_t axis_simd_full_;
-    size_t axis_simd_tail_;
+    int axis_simd_tail_;
     size_t n_loops_;
     size_t loop_tail_;
     size_t src_next_vreg_stride_;
@@ -1544,7 +1544,7 @@ struct jit_softmax_strided_kernel_t : jit_softmax_kernel_base_t,
                     tmp_vmm_injector, this->r14, this->r15, this->r13,
                     preserve_gpr, preserve_vmm,
                     PARAM_OFF(post_ops_binary_rhs_arg_vec), PARAM_OFF(dst_orig),
-                    dst_d_, static_cast<int>(axis_simd_tail_), tail_opmask,
+                    dst_d_, axis_simd_tail_, tail_opmask,
                     use_exact_tail_scalar_bcast};
 
             const binary_injector::static_params_t bsp {
@@ -1594,7 +1594,7 @@ struct jit_softmax_strided_kernel_t : jit_softmax_kernel_base_t,
         // different pieces from the dense version.
         , axis_stride_(pd_->axis_stride())
         , axis_simd_full_(axis_stride_ / simd_w_)
-        , axis_simd_tail_(axis_stride_ % simd_w_) {
+        , axis_simd_tail_(static_cast<int>(axis_stride_ % simd_w_)) {
 
         // Scratchpad size is limited to a single simd_w, thus, no unrolling
         // for such cases.
